requirements.txt
streamlit
altair
pdf2image
PyPDF2
openai
python-dotenv
